refactor(ViewController): 템플릿을 활용하여 토픽 바텀시트 리팩토링(#76)
